Move sandbox logging to orchestrator
